+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+# 3.7.0
+
+* fixing AssetsBucket when a bucket name is provided
+* Adding AuroraMysql construct
+
 # 3.6.2
 
 * updating asg allowed instance types